# src/main/java/com/n19dccn112/model/entity/OrderDetail.java
package com.n19dccn112.model.entity;

import com.n19dccn112.model.key.OrderDetailId;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import javax.persistence.*;

@Entity
@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Table(name = "order_detail")
public class OrderDetail {
    @EmbeddedId
    private OrderDetailId orderDetailId;

    @Column(name = "amount")
    private int amount;
}